Merge branch 'treesitter' of github.com:Shrimpram/vim-stella
Add Treesitter highlight groups

Fix TS Text highlight

Set it to lik to Normal instead of "NONE"

Change TS Error to Normal

TS Error isn't for error codes but rather for when *Treesitter's parser*
can't detect which syntax to use for specific text. This makes it so
that TS falls back to the default text syntax when the syntax parser
returns an error.